arduino/requirements.txt
pyserial==2.7
requests==2.7.0
django-dotenv==1.3.0
django-getenv==1.3.1
